refactor: MemberType enum 위치 변경
